app/requirements.txt
langchain~=0.3.26
python-dotenv~=1.1.0
transformers~=4.52.4
fastapi[all]~=0.115.14
mongoengine==0.20.0
pydantic~=2.11.7

uvicorn~=0.35.0
